reorganize location of story file
